cxl/acpi: Fix setup of memory resource in cxl_acpi_set_cache_size()
In order to compare the resource against the HMAT memory target,
the resource needs to be memory type. Change the DEFINE_RES()
macro to DEFINE_RES_MEM() in order to set the correct resource type.
This causes a regression for platforms with the extended linear cache
enabled as the comparison always fails and the cache size is not set.
User visible impact is that when 'cxl list' is issued, a CXL region
with extended linear cache support will only report half the size
of the actual size. And this also breaks MCE reporting of the memory
region due to incorrect offset calculation for the memory.

Fixes: 12b3d697c812 ("cxl: Remove core/acpi.c and cxl core dependency on ACPI")
Cc: stable@kernel.org
Reviewed-by: Gregory Price <gourry@gourry.net>
Reviewed-by: Alison Schofield <alison.schofield@intel.com>
Signed-off-by: Dave Jiang <dave.jiang@intel.com>